testing/web-platform/mozilla/tests/binast/insecure.html

<!DOCTYPE html>
<title>Check we dont't load BinAST over an insecure connection</title>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    setup({allow_uncaught_exception: true});

    var hadError = false;
    function setError() { hadError = true; }

    window.addEventListener("error", setError);

    const test_load = async_test("Check we can't load BinAST over HTTP");
    window.addEventListener("load", test_load.step_func_done(ev => {
      assert_equals(hadError, false, "Didn't expect an error event");
      assert_equals(binASTLoaded, false, "Didn't expect to load BinAST version");
    }));

</script>
<script src="./serve.py?name=small" onerror="setError()"></script>
